Moved selectedElementList back to the vuex store and refactored the Queue component to use the vuex store
