+++ src/components/HomeSidebar.vue
@@ -9,9 +9,6 @@
     </section>
     <!-- </v-form> -->
     <section>
-      <h1 class="headline purple--text text--accent-2">Selected Elements</h1>
-      <hr>
-      <Queue :listToRender="selectedElementList"/>
       <v-select
         v-model="selectedChildren"
         :items="Object.keys(componentMap).filter(comp => comp !== 'App')"
@@ -21,6 +18,9 @@
         hint="Select child components"
         persistent-hint
       ></v-select>
+      <h1 class="headline purple--text text--accent-2">Selected Elements</h1>
+      <hr>
+      <Queue :listToRender="selectedElementList"/>
     </section>
 
     <BaseButton :componentName="componentName" name="add component" icon="add_circle"></BaseButton>
@@ -73,7 +73,7 @@ export default {
 <style>
 .sidebar {
   display: grid;
-  grid-template-rows: 0.5fr 0.5 300px 4fr 1fr;
+  grid-template-rows: 0.5fr 0.5fr 0.5fr 4fr 1fr;
 }
 
 /* .sidebar {
