+++ src/router.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import Router from 'vue-router';
-import Home from './views/Home.vue';
+import Home from './views/HomeView.vue';
 
 Vue.use(Router);
 
@@ -20,7 +20,7 @@ export default new Router({
       // this generates a separate chunk (about.[hash].js) for this route
       // which is lazy-loaded when the route is visited.
       component: () =>
-        import(/* webpackChunkName: "about" */ './views/Tree.vue')
+        import(/* webpackChunkName: "about" */ './views/TreeView.vue')
     }
   ]
 });
