+++ src/store/actions.js
@@ -6,6 +6,11 @@ const actions = {
     if (!state.componentMap[name]) {
       commit(types.ADD_TO_COMPONENT_MAP, payload);
     }
+  },
+  [types.setSelectedElementList]: ({ state, commit }, payload) => {
+    if (payload) {
+      commit(types.SET_SELECTED_ELEMENT_LIST, payload);
+    }
   }
 };
