+++ src/store/types.js
@@ -1,5 +1,7 @@
 //Mutations
 export const ADD_TO_COMPONENT_MAP = 'ADD_TO_COMPONENT_MAP';
+export const SET_SELECTED_ELEMENT_LIST = 'SET_SELECTED_ELEMENT_LIST';
 
 //Actions
 export const registerComponent = 'registerComponent';
+export const setSelectedElementList = 'setSelectedElementList';
